refactor: 댓글방 상태 도메인 설계 변경
