--- backend/src/main/java/com/zzang/chongdae/comment/service/dto/CommentRoomStatusResponse.java
@@ -1,6 +1,0 @@
-package com.zzang.chongdae.comment.service.dto;
-
-import com.zzang.chongdae.offering.domain.status.CommentRoomStatus;
-
-public record CommentRoomStatusResponse(CommentRoomStatus updatedStatus) {
-}
