--- backend/src/main/java/com/zzang/chongdae/offering/domain/status/CommentRoomCondition.java
@@ -1,8 +1,0 @@
-package com.zzang.chongdae.offering.domain.status;
-
-public abstract class CommentRoomCondition {
-
-    public abstract CommentRoomCondition nextCondition();
-
-    public abstract CommentRoomStatus status();
-}
