--- backend/src/main/java/com/zzang/chongdae/offering/domain/status/Trading.java
@@ -1,17 +1,0 @@
-package com.zzang.chongdae.offering.domain.status;
-
-import static com.zzang.chongdae.offering.domain.status.CommentRoomStatus.DONE;
-import static com.zzang.chongdae.offering.domain.status.CommentRoomStatus.TRADING;
-
-public class Trading extends CommentRoomCondition {
-
-    @Override
-    public CommentRoomCondition nextCondition() {
-        return DONE.getCondition();
-    }
-
-    @Override
-    public CommentRoomStatus status() {
-        return TRADING;
-    }
-}
