+++ backend/src/main/java/com/zzang/chongdae/offering/service/dto/CommentRoomStatusResponse.java
@@ -1,0 +1,6 @@
+package com.zzang.chongdae.offering.service.dto;
+
+import com.zzang.chongdae.offering.domain.CommentRoomStatus;
+
+public record CommentRoomStatusResponse(CommentRoomStatus updatedStatus) {
+}
