+++ backend/src/test/java/com/zzang/chongdae/global/domain/OfferingFixture.java
@@ -1,7 +1,7 @@
 package com.zzang.chongdae.global.domain;
 
 import com.zzang.chongdae.member.repository.entity.MemberEntity;
-import com.zzang.chongdae.offering.domain.status.CommentRoomStatus;
+import com.zzang.chongdae.offering.domain.CommentRoomStatus;
 import com.zzang.chongdae.offering.repository.OfferingRepository;
 import com.zzang.chongdae.offering.repository.entity.OfferingEntity;
 import java.time.LocalDateTime;
